requirements.txt
django
djangorestframework
pillow
django-imagekit